Revert "ref:move configuration providers to right places"
This reverts commit 594108ecf3614f9e361df0bccb704a9d1fa02050.
