--- Campus.Infrastructure.Configuration/Campus.Infrastructure.Configuration.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Src\Campus.Services.Interfaces\Campus.Services.Interfaces.csproj" />
-  </ItemGroup>
-
-</Project>
