+++ Campus.Master.API.sln
@@ -3,27 +3,25 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29709.97
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Master.API", "Src\Campus.Master.API\Campus.Master.API.csproj", "{B7A332F4-936D-4572-82D1-088B40B3CC2C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Master.API", "Src\Campus.Master.API\Campus.Master.API.csproj", "{B7A332F4-936D-4572-82D1-088B40B3CC2C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{BE04D635-A676-4D41-9BA0-7A95774E425F}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{6DF4B688-FF8F-4CFD-961F-9AA989BD03BD}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Application", "Application", "{28CA258B-8F72-47EB-BAC8-64C6D743487D}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{28CA258B-8F72-47EB-BAC8-64C6D743487D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Domain.Core", "Src\Campus.Domain.Core\Campus.Domain.Core.csproj", "{138BBAAC-F414-4AB0-B38C-79F718620107}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Domain.Core", "Src\Campus.Domain.Core\Campus.Domain.Core.csproj", "{138BBAAC-F414-4AB0-B38C-79F718620107}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Services.Interfaces", "Src\Campus.Services.Interfaces\Campus.Services.Interfaces.csproj", "{278B45F4-A347-4540-80A0-1B04A93340F0}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Services.Interfaces", "Src\Campus.Services.Interfaces\Campus.Services.Interfaces.csproj", "{278B45F4-A347-4540-80A0-1B04A93340F0}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Infrastructure.Data.EntityFrameworkCore", "Src\Campus.Infrastructure.Data.EntityFrameworkCore\Campus.Infrastructure.Data.EntityFrameworkCore.csproj", "{0E3C28A9-E86F-4265-964C-626288CE4A2E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Infrastructure.Data.EntityFrameworkCore", "Src\Campus.Infrastructure.Data.EntityFrameworkCore\Campus.Infrastructure.Data.EntityFrameworkCore.csproj", "{0E3C28A9-E86F-4265-964C-626288CE4A2E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test", "Test", "{0FB95748-D31C-47E8-8B93-DBC1796234CA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Master.IntegrationTests", "Test\Campus.Master.IntegrationTests\Campus.Master.IntegrationTests.csproj", "{DA9E68CB-AF98-4921-88EC-FB5E038EB5FE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Master.IntegrationTests", "Test\Campus.Master.IntegrationTests\Campus.Master.IntegrationTests.csproj", "{DA9E68CB-AF98-4921-88EC-FB5E038EB5FE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Campus.Services.Implementation", "Src\Campus.Services\Campus.Services.Implementation.csproj", "{6725C9B8-3804-4B4F-8314-0C9EE4CC5829}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Infrastructure.Configuration", "Campus.Infrastructure.Configuration\Campus.Infrastructure.Configuration.csproj", "{725710CA-3340-4718-9FA0-0E6825E50F0E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Campus.Services", "Src\Campus.Services\Campus.Services.csproj", "{6725C9B8-3804-4B4F-8314-0C9EE4CC5829}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -55,23 +53,18 @@ Global
 		{6725C9B8-3804-4B4F-8314-0C9EE4CC5829}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6725C9B8-3804-4B4F-8314-0C9EE4CC5829}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6725C9B8-3804-4B4F-8314-0C9EE4CC5829}.Release|Any CPU.Build.0 = Release|Any CPU
-		{725710CA-3340-4718-9FA0-0E6825E50F0E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{725710CA-3340-4718-9FA0-0E6825E50F0E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{725710CA-3340-4718-9FA0-0E6825E50F0E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{725710CA-3340-4718-9FA0-0E6825E50F0E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BEE640B1-4E3D-4EF7-9E2F-6B7E186BFCC2}
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{138BBAAC-F414-4AB0-B38C-79F718620107} = {BE04D635-A676-4D41-9BA0-7A95774E425F}
 		{278B45F4-A347-4540-80A0-1B04A93340F0} = {28CA258B-8F72-47EB-BAC8-64C6D743487D}
 		{0E3C28A9-E86F-4265-964C-626288CE4A2E} = {6DF4B688-FF8F-4CFD-961F-9AA989BD03BD}
 		{DA9E68CB-AF98-4921-88EC-FB5E038EB5FE} = {0FB95748-D31C-47E8-8B93-DBC1796234CA}
-		{6725C9B8-3804-4B4F-8314-0C9EE4CC5829} = {6DF4B688-FF8F-4CFD-961F-9AA989BD03BD}
-		{725710CA-3340-4718-9FA0-0E6825E50F0E} = {6DF4B688-FF8F-4CFD-961F-9AA989BD03BD}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {BEE640B1-4E3D-4EF7-9E2F-6B7E186BFCC2}
+		{6725C9B8-3804-4B4F-8314-0C9EE4CC5829} = {28CA258B-8F72-47EB-BAC8-64C6D743487D}
 	EndGlobalSection
 EndGlobal
