+++ Src/Campus.Domain.Core/Interfaces/IConfigurationProviderFactory.cs
@@ -1,4 +1,4 @@
-namespace Campus.Services.Interfaces.Interfaces.Configuration
+namespace Campus.Domain.Core.Interfaces
 {
     public interface IConfigurationProviderFactory
     {
