+++ Src/Campus.Master.API/Campus.Master.API.csproj
@@ -31,10 +31,9 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\Campus.Infrastructure.Configuration\Campus.Infrastructure.Configuration.csproj" />
       <ProjectReference Include="..\Campus.Infrastructure.Data.EntityFrameworkCore\Campus.Infrastructure.Data.EntityFrameworkCore.csproj" />
       <ProjectReference Include="..\Campus.Services.Interfaces\Campus.Services.Interfaces.csproj" />
-      <ProjectReference Include="..\Campus.Services\Campus.Services.Implementation.csproj" />
+      <ProjectReference Include="..\Campus.Services\Campus.Services.csproj" />
     </ItemGroup>
 
     <ItemGroup>
