+++ Src/Campus.Master.API/Helpers/Implementations/AppSettingsConfigurationProvider.cs
@@ -1,8 +1,8 @@
 using System;
 using Microsoft.Extensions.Configuration;
-using ICampusConfigurationProvider = Campus.Services.Interfaces.Interfaces.Configuration.IConfigurationProvider;
+using ICampusConfigurationProvider = Campus.Domain.Core.Interfaces.IConfigurationProvider;
 
-namespace Campus.Infrastructure.Configuration.Implementation
+namespace Campus.Master.API.Helpers.Implementations
 {
     public class AppSettingsConfigurationProvider : ICampusConfigurationProvider
     {
