+++ Src/Campus.Master.API/Helpers/Implementations/ConfigurationProviderFactory.cs
@@ -1,8 +1,8 @@
-using Campus.Services.Interfaces.Interfaces.Configuration;
+using Campus.Domain.Core.Interfaces;
 using Microsoft.Extensions.Configuration;
-using ICampusConfigurationProvider = Campus.Services.Interfaces.Interfaces.Configuration.IConfigurationProvider;
+using ICampusConfigurationProvider = Campus.Domain.Core.Interfaces.IConfigurationProvider;
 
-namespace Campus.Infrastructure.Configuration.Implementation
+namespace Campus.Master.API.Helpers.Implementations
 {
     public class ConfigurationProviderFactory : IConfigurationProviderFactory
     {
