+++ Src/Campus.Services/Campus.Services.csproj
