+++ Src/Campus.Services/Core/ClassroomService.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Campus.Domain.Core.Models;
@@ -6,7 +7,7 @@ using Campus.Services.Interfaces.DTO.Classroom;
 using Campus.Services.Interfaces.Interfaces;
 using Microsoft.EntityFrameworkCore;
 
-namespace Campus.Services.Implementation.Core
+namespace Campus.Services.Core
 {
     public class ClassroomService : IClassroomService
     {
