+++ Src/Campus.Services/Core/FakeDetectorService.cs
@@ -1,18 +1,17 @@
-﻿using System;
-using System.Net.Http;
+﻿using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading.Tasks;
 using Campus.Services.Interfaces.DTO.Auxiliary;
 using Campus.Services.Interfaces.Interfaces;
-using ICampusConfigurationProvider = Campus.Services.Interfaces.Interfaces.Configuration.IConfigurationProvider;
+using Microsoft.Extensions.Configuration;
 
-namespace Campus.Services.Implementation.Core
+namespace Campus.Services.Core
 {
     class FakeDetectorService : IFakeDetectorService
     {
-        private readonly ICampusConfigurationProvider _configuration;
+        private readonly IConfiguration _configuration;
 
-        public FakeDetectorService(ICampusConfigurationProvider configuration)
+        public FakeDetectorService(IConfiguration configuration)
         {
             _configuration = configuration;
         }
@@ -21,18 +20,21 @@ namespace Campus.Services.Implementation.Core
         {
             var fakeDetectorRequest = new HttpClient();
 
-            fakeDetectorRequest.DefaultRequestHeaders.Add("x-functions-key",
-                _configuration.GetConfigurationValue("FakeDetector:ApiKey", Convert.ToString));
+            var fakeDetectorSettings = _configuration
+                .GetSection("FakeDetector");
 
+            fakeDetectorRequest
+                .DefaultRequestHeaders
+                .Add("x-functions-key", fakeDetectorSettings.GetValue<string>("ApiKey"));
 
+            
             var byteArrayContent = new ByteArrayContent(imageData);
             byteArrayContent.Headers.ContentType = MediaTypeHeaderValue.Parse(attributes.ContentType);
 
             var response = await fakeDetectorRequest.PostAsync(
-                _configuration.GetConfigurationValue("FakeDetector:Uri", Convert.ToString),
-                new MultipartFormDataContent
+                fakeDetectorSettings.GetValue<string>("Uri"), new MultipartFormDataContent
                 {
-                    {byteArrayContent, "\"file\"", attributes.FileName}
+                    { byteArrayContent, "\"file\"", attributes.FileName }
                 });
 
             return await response.Content.ReadAsStringAsync();
